+++ sfxc/src/input_node_tasklet.cc
@@ -206,7 +206,7 @@ set_parameters(const Input_node_parameters &input_node_param,
   int nr_ffts = Control_parameters::
                 nr_ffts_per_integration_slice((int)input_node_param.integr_time.get_time_usec(),
                                               sample_rate, input_node_param.fft_size);
-  size_slice = nr_ffts * input_node_param.fft_size;
+  size_slice = (int64_t)nr_ffts * input_node_param.fft_size;
 }
